:truck: Rename and migrate reader functions to io
