+++ kf_lib_data_ingest/common/misc.py
@@ -2,14 +2,11 @@ import ast
 import datetime
 import importlib
 import inspect
-import json
 import os
 import re
 import time
 from itertools import tee
 
-import jsonpickle
-import yaml
 from pandas import isnull
 
 from kf_lib_data_ingest.common.type_safety import (
@@ -49,58 +46,6 @@ def import_module_from_file(filepath):
     imported_module = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(imported_module)
     return imported_module
-
-
-def read_yaml(filepath):
-    with open(filepath, "r") as yaml_file:
-        return yaml.load(yaml_file, Loader=yaml.FullLoader)
-
-
-def read_json(filepath, default=None, use_jsonpickle=True):
-    """
-    Read JSON file into Python dict. If default is not None and the file
-    does not exist, then return default.
-
-    :param filepath: path to JSON file
-    :type filepath: str
-    :param default: default return value if file not found, defaults to None
-    :type default: any, optional
-    :param use_jsonpickle: pickle JSON-incompatible types, defaults to True
-    :type use_jsonpickle: bool, optional
-    :return: your data
-    :rtype: dict
-    """
-    if (default is not None) and (not os.path.isfile(filepath)):
-        return default
-
-    with open(filepath, "r") as json_file:
-        if use_jsonpickle:
-            json_str = json_file.read()
-            return jsonpickle.decode(json_str, keys=True)
-        else:
-            return json.load(json_file)
-
-
-def write_json(data, filepath, use_jsonpickle=True, **kwargs):
-    r"""
-    Write Python dict to JSON file.
-
-    :param data: your data
-    :type data: dict
-    :param filepath: where to write your JSON file
-    :type filepath: str
-    :param use_jsonpickle: pickle JSON-incompatible types, defaults to True
-    :type use_jsonpickle: bool, optional
-    :param \**kwargs: keyword arguments to pass to json.dump
-    """
-    if "indent" not in kwargs:
-        kwargs["indent"] = 4
-    if "sort_keys" not in kwargs:
-        kwargs["sort_keys"] = True
-    with open(filepath, "w") as json_file:
-        if use_jsonpickle:
-            data = json.loads(jsonpickle.encode(data, keys=True))
-        json.dump(data, json_file, **kwargs)
 
 
 def iterate_pairwise(iterable):
